snp count depending on allele

+++ limix_reader/util/ndict.py
@@ -5,4 +5,4 @@ class ndict(dict):
     def __getitem__(self, key):
         if isscalar(key):
             return super(ndict, self).__getitem__(key)
-        return asarray([self[k] for k in key])
+        return asarray([self[k] for k in key], dtype=int)
